server/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
groq==0.11.0
httpx==0.27.0
python-dotenv==1.0.1
PyYAML==6.0.2